requirements.txt
apache-airflow==2.9.2
apache-airflow-providers-mysql==5.6.2
pendulum==3.0.0
duckdb==1.0.0
